backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.1
numpy==1.26.1
scikit-learn==1.3.2
matplotlib==3.8.1
seaborn==0.13.0

# Geospatial
geopandas==0.14.0
folium==0.14.0
shapely==2.0.2
pyproj==3.6.1

# Weather API
requests==2.31.0
aiohttp==3.8.6

# AI/ML
openai==1.2.4
langchain==0.0.335
transformers==4.35.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
joblib==1.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1